Initial checkin moving utilities to kms_client_provider
