+++ src/cpio/client_providers/kms_client_provider/azure/azure_kms_client_provider.h
@@ -26,9 +26,8 @@
 #include "src/azure/attestation/src/attestation.h"
 #include "src/core/interface/async_context.h"
 #include "src/cpio/client_providers/interface/kms_client_provider_interface.h"
-#include "src/cpio/client_providers/private_key_fetcher_provider/azure/azure_private_key_fetcher_provider_utils.h"
-#include "src/cpio/client_providers/private_key_fetcher_provider/private_key_fetcher_provider_utils.h"
 #include "src/public/core/interface/execution_result.h"
+#include "azure_kms_client_provider_utils.h"
 
 namespace google::scp::cpio::client_providers {
